fix: sync navbar price and price blocks in home page (#935)
