+++ app/Http/Livewire/NavbarSettings.php
@@ -30,7 +30,6 @@ final class NavbarSettings extends Component
 
         if ($originalCurrency !== $newCurrency) {
             $this->emit('currencyChanged', $newCurrency);
-            $this->emit('refreshNetworkStatusBlock');
         }
 
         if ($originalTheme !== $newTheme) {
